requirements.txt
boto3==1.17.79
botocore==1.20.79
python-decouple==3.4
requests==2.22.0
